Fix: Suppress unwanted output on stderr
Like it already said in the comments: The `cd` call might fail and thats ok, since it's ignored. However, the command `cd` sometimes prints stuff to stderr, for example if the current user lack permission to read the directory. This is pretty annoying in an environment where the .multirust config folder (hence the overrides) is shared between multiple users. So since we want to ignore every error from `cd` anyway, I piped the stderr to /dev/null to silence the command.